make-release-build.sh: add ndk-build and make script run on any system
The ant.properties file is only needed when making a final, signed build.

+++ make-release-build.sh
@@ -2,7 +2,11 @@
 
 set -e
 
-. ~/.android/bashrc
+if [ -e ~/.android/bashrc ]; then
+    . ~/.android/bashrc
+else
+    echo "No ~/.android/bashrc found, 'android' and 'ndk-build' must be in PATH"
+fi
 
 projectroot=`pwd`
 projectname=`sed -n 's,.*name="app_name">\(.*\)<.*,\1,p' app/res/values/strings.xml`
@@ -18,7 +22,15 @@ git reset --hard
 git clean -fdx
 git submodule update --init --recursive
 
-cp ~/.android/ant.properties $projectroot/app/
+echo "Running ndk-build:"
+ndk-build -C $projectroot/external/F5Android
+
+if [ -e ~/.android/ant.properties ]; then
+    cp ~/.android/ant.properties $projectroot/app/
+else
+    echo "skipping release ant.properties"
+fi
+
 ./setup-ant.sh
 
 cd $projectroot/app/
